Change Tach input... Engine now starts
The tach input was previously sourced from the the old ignition driver,
which missed every 3rd or every other trigger event while the starter
was cranking.  This prevented the ECU from producing accurate dwell
times while cranking.  We now source timing directly from the hall
effect (a wiring change in the harness), which produces the expected,
consistent square wave.  The signal is inverted, so we're back to rising
edge trigger.