Fixed type mismatch of rpm var in tweaker class
The rpm variable is passed by reference to the boss object (FuelTweaker
class).  The FuelTweaker class was expecting a reference to a signed
int, but it was given a reference to an unsigned int.  So boss was
always seeing an erroneous rpm value, which totally buggered where the
changes to the local map were going.